+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.0.1] - 2021-04-03
+## [3.0.2] - 2021-04-03
 ### Changed
 - Change error message
+- Optimize SourceFileContainer
 
 ## [3.0.0] - 2021-03-21
 ### Added